byol_finetuning/requirements.txt
matplotlib==3.3.2
numpy==1.19.1
opencv_python_headless==4.4.0.46
pandas==1.1.5
Pillow==8.4.0
pytorch_lightning==1.0.7
scikit_learn
seaborn==0.11.1
torch==1.10.0
torchvision==0.11.1
